feat: Implement secure authentication, migrate IDs to Long, and expand tests
This commit refactors the authentication flow using MVVM and BCrypt, migrates all database entities to use Long IDs, and adds comprehensive integration tests.

- **Authentication and Security:**
    - Implement `LoginViewModel` and `LoginUiState` to manage login logic and UI state.
    - Create `AuthRepository` and `PasswordHasher` utility using BCrypt for secure password handling.
    - Update `RegisterScreen` with strong password validation, duplicate email checks, and password hashing before storage.
    - Refactor `LoginScreen` to observe ViewModel state and remove direct database calls.
    - Remove insecure `login` query from `UserDao`.

- **Database and Data Models:**
    - Migrate primary and foreign keys from `Int` to `Long` across `User`, `Car`, `Poster`, and `PosterImage` entities and DAOs.
    - Add `Index` annotations to `User` (email) and `Poster` (carId) for performance and uniqueness constraints.
    - Define Foreign Keys with `onDelete = CASCADE` in `Poster` and `PosterImage`.
    - Implement `equals` and `hashCode` for `Poster` and `PosterImage` entities.
    - Refactor database instantiation into a singleton `DatabaseBuilder`.
    - Increment `AppDatabase` version to 12.

- **UI and Refactoring:**
    - Update `MainActivity`, `AnunciosScreen`, `CarDetailScreen`, and `PosterAdapter` to handle `Long` identifiers.
    - Add mapping logic in `CarDetailScreen` to display localized strings for fuel types, categories, and accessories.
    - Remove unused `NavGraph.kt` file.

- **Testing and Configuration:**
    - Add dependencies for `jbcrypt` and Android Test Orchestrator in `build.gradle.kts`.
    - Create integration tests: `LoginIntegrationTest` for user flows and `PosterCarDaoIntegrationTest` for relationship integrity.
    - Configure test options to enable resource inclusion and disable animations.

+++ app/src/main/java/com/example/luxcar/data/model/Car.kt
@@ -6,7 +6,7 @@ import androidx.room.TypeConverter
 
 @Entity(tableName = "cars")
 data class Car (
-    @PrimaryKey(autoGenerate = true) val id: Int=0,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0L,
     val marca: String,
     val modelo: String,
     val cor: String,
@@ -17,11 +17,11 @@ data class Car (
     val acessorios: List<String>
 )
 
-class Converters{
+class Converters {
     @TypeConverter
     fun fromList(list: List<String>): String = list.joinToString(",")
 
     @TypeConverter
     fun toList(data: String): List<String> =
-            if(data.isEmpty()) emptyList() else data.split(",")
+        if(data.isEmpty()) emptyList() else data.split(",")
 }
